+++ contracts/__init__.py
@@ -8,6 +8,7 @@ from .i_event_bus import IEventBus
 from .i_capability_registry import ICapabilityRegistry
 from .igraph_builder import IGraphBuilder
 from .igraph_query import IGraphQuery
+from .snapshotable import ISnapshotable
 
 __all__ = [
     "IService",
@@ -16,4 +17,5 @@ __all__ = [
     "ICapabilityRegistry",
     "IGraphBuilder",
     "IGraphQuery",
+    "ISnapshotable",
 ]