+++ infrastructure/__init__.py
@@ -3,10 +3,12 @@ from .container import DependencyContainer
 from .eventbus import InMemoryEventBus
 from .graph_builder import InMemoryGraphBuilder
 from .config_loader import ConfigLoader
+from .snapshot_store import SnapshotStore
 
 __all__ = [
     "DependencyContainer",
     "InMemoryEventBus",
     "InMemoryGraphBuilder",
     "ConfigLoader",
+    "SnapshotStore",
 ]